backend: fix logout flow typings

+++ backend/src/core/services/auth/logout.service.ts
@@ -1,13 +1,16 @@
 import { Injectable } from '@nestjs/common';
-import { LogoutDto, LogoutSuccessDto } from '@shared/dto';
+import { LogoutSuccessDto } from '@shared/dto';
+import { InvalidParamsException } from '@src/core/exceptions/invalid-params.exception';
 import { RefreshSessionDao } from '@src/data/dao/refresh-session/refresh-session.dao';
 
 @Injectable()
 export class LogoutService {
   constructor(private readonly refreshSessionDao: RefreshSessionDao) {}
 
-  async logout(logoutDto: LogoutDto): Promise<LogoutSuccessDto> {
-    const { refreshToken } = logoutDto;
+  async logout(refreshToken: string): Promise<LogoutSuccessDto> {
+    if (!refreshToken) {
+      throw new InvalidParamsException('Refresh token not provided');
+    }
 
     await this.refreshSessionDao.deleteAll({ refreshToken });
